Split Qwen35 oracle prefill trace
